refactor(web): split counter history route into its own file
qlty flagged counters.ts for high total complexity. Following the
existing sfxCategoryMutations/sfxTriggerMutations/sfxFileMutations
split pattern, moved GET /counters/:id/history into a standalone
counterHistory.ts router, mounted alongside countersRouter.
