+++ src/web/server.test.ts
@@ -96,6 +96,7 @@ vi.mock('./routes/sfxPublic', () => ({ default: emptyRouter() }));
 vi.mock('./routes/streams', () => ({ default: markerRouter('streams') }));
 vi.mock('./routes/commands', () => ({ default: emptyRouter() }));
 vi.mock('./routes/counters', () => ({ default: emptyRouter() }));
+vi.mock('./routes/counterHistory', () => ({ default: emptyRouter() }));
 vi.mock('./routes/commandMonitor', () => ({ default: emptyRouter() }));
 vi.mock('./routes/streamdeck', () => ({ default: emptyRouter() }));
 vi.mock('./routes/streamdeckKeys', () => ({ default: emptyRouter() }));
