+++ src/web/server.ts
@@ -24,6 +24,7 @@ import sfxPublicRouter from './routes/sfxPublic';
 import streamsRouter from './routes/streams';
 import commandsRouter from './routes/commands';
 import countersRouter from './routes/counters';
+import counterHistoryRouter from './routes/counterHistory';
 import commandMonitorRouter from './routes/commandMonitor';
 import streamdeckRouter from './routes/streamdeck';
 import streamdeckKeysRouter from './routes/streamdeckKeys';
@@ -192,6 +193,7 @@ rootAuthedRouter.use(requireAuth);
 rootAuthedRouter.use(sfxRouter);
 rootAuthedRouter.use(commandsRouter);
 rootAuthedRouter.use(countersRouter);
+rootAuthedRouter.use(counterHistoryRouter);
 rootAuthedRouter.use(commandMonitorRouter);
 rootAuthedRouter.use(rootGuildRouter);
 app.use('/', rootAuthedRouter);
